requirements.txt
mock
python-magic==0.4.12